Added some functions/properties (read description)
You can now create Proton3DObjects (maybe importing them as well?) with the `mirror` parameter set to `true` in order to have real time reflections with the other objects in the scene at the time the object was created.

You can now also call [scene].setEnvironmentMap with a URL to an HDR file to get a environment map for all objects in a scene. This works even better if you find some way to bake all your textures. HOWEVER on my laptop there is slow performance. That is all. Have a good day/night/whatever!